requirements.txt
Flask==3.1.0
pandas==2.2.3
scikit-learn==1.6.1
gunicorn==21.2.0
numpy>=1.26.0
python-dotenv==1.0.1
